[Hong]<test>Add myrecipe-testcode

+++ src/controllers/favorite.ts
@@ -2,15 +2,19 @@ import { Request, Response, NextFunction } from "express";
 import Favorite from "../models/favorite";
 import { IFavorite } from "../interfaces/favorite";
 import myrecipe from "../models/myrecipe";
+import Mystore from "../models/store";
 
 //마이레시피에 좋아요 누르기
 const postlike = async (req: Request, res: Response) => {
     const userId = String(res.locals.user._id);
     const nickname = String(res.locals.user.nickname);
-    const {  myrecipeId  } = req.params;
+    const { myrecipeId } = req.params;
     try {
         const existLike = await Favorite.findOne({ _id: userId, myrecipeId: myrecipeId, nickname: nickname });
-
+        const Myrecipe = await myrecipe.findById({ _id: myrecipeId });
+        if (!Myrecipe) {
+            return res.json({ result: false, msg: "존재하지 않는 값입니다." });
+        }
         if (existLike) {
             return res.json({ result: false, msg: "이미 좋아요를 누르셨습니다." });
         } else {
@@ -19,17 +23,14 @@ const postlike = async (req: Request, res: Response) => {
                 myrecipeId,
                 nickname
             });
-            res.json({ result: true, favorite });
+            let num: number = Myrecipe.favorite_count;
+            await myrecipe.findOneAndUpdate({ _id: myrecipeId }, { $set: { favorite_count: ++num } });
+            return res.json({ result: true, favorite });
         }
-        // const Myrecipe = await myrecipe.findById({ _id: myrecipeId})
-
-        // // let num: number = Myrecipe.favorite_count;
-
-        // await myrecipe.findOneAndUpdate({ _id: myrecipeId }, {$set: {favorite_count: ++num}})
-        // return res.json({ result: true })
     } catch (err) {
         console.log(err);
         res.json({ result: false });
+        return;
     }
 };
 
@@ -39,9 +40,11 @@ const getAlluser = async (req: Request, res: Response) => {
     try {
         const getUser: Array<IFavorite> = await Favorite.find({ myrecipeId });
         res.json({ result: true, getUser });
+        return;
     } catch (err) {
         console.log(err);
         res.json({ result: false });
+        return;
     }
 };
 
@@ -52,15 +55,23 @@ const deletelike = async (req: Request, res: Response) => {
     const nickname = String(res.locals.user.nickname);
     try {
         const findUser = await Favorite.findOne({ userId, myrecipeId, nickname });
+        const Myrecipe = await myrecipe.findById({ _id: myrecipeId });
+        if (!Myrecipe) {
+            res.json({ result: false, msg: "존재하지 않는 값입니다." });
+            return;
+        }
         if (!findUser) {
             return res.json({ result: false, msg: "좋아요를 누르지 않았습니다." });
         } else {
             await Favorite.findByIdAndDelete(findUser);
+            let num: number = Myrecipe.favorite_count;
+            await myrecipe.findOneAndUpdate({ _id: myrecipeId }, { $set: { favorite_count: --num } });
             return res.json({ result: true });
         }
     } catch (err) {
         console.log(err);
         res.json({ result: false });
+        return;
     }
 };
 
@@ -69,11 +80,12 @@ const getMyrecipe = async (req: Request, res: Response) => {
     const userId = String(res.locals.user._id);
     try {
         const getMyrecipe: Array<IFavorite> = await Favorite.find({ userId });
-       
+
         res.json({ result: true, getMyrecipe });
     } catch (err) {
         console.log(err);
         res.json({ result: false });
+        return;
     }
 };
 
@@ -82,33 +94,38 @@ const getMyrecipe = async (req: Request, res: Response) => {
 const postStorelike = async (req: Request, res: Response) => {
     const userId = String(res.locals.user._id);
     const nickname = String(res.locals.user.nickname);
-    const { 
-        mystoreId,
-     } = req.params;
+    const { MystoreId } = req.params;
     try {
-        const existLike = await Favorite.findOne({ _id: userId, mystoreId:mystoreId, nickname: nickname });
-
+        const existLike = await Favorite.findOne({ _id: userId, MystoreId: MystoreId, nickname: nickname });
+        const Store = await Mystore.findById({ _id: MystoreId });
+        if (!Store) {
+            res.json({ result: false, msg: "존재하지 않는 값입니다." });
+            return;
+        }
         if (existLike) {
             return res.json({ result: false, msg: "이미 좋아요를 누르셨습니다." });
         } else {
             const favorite = await Favorite.create({
                 userId,
-                mystoreId,
+                MystoreId,
                 nickname
             });
-            res.json({ result: true, favorite });
+            let num: number = Store.favorite_count;
+            await myrecipe.findOneAndUpdate({ _id: MystoreId }, { $set: { favorite_count: ++num } });
+            return res.json({ result: true, favorite });
         }
     } catch (err) {
         console.log(err);
         res.json({ result: false });
+        return;
     }
 };
 
 // 마이스토어 좋아요 누른 사람 조회
 const getAllstoreuser = async (req: Request, res: Response) => {
-    const { mystoreId } = req.params;
+    const { MystoreId } = req.params;
     try {
-        const getUser: Array<IFavorite> = await Favorite.find({ mystoreId });
+        const getUser: Array<IFavorite> = await Favorite.find({ MystoreId });
         res.json({ result: true, getUser });
     } catch (err) {
         console.log(err);
@@ -118,15 +135,22 @@ const getAllstoreuser = async (req: Request, res: Response) => {
 
 // 마이스토어 좋아요 삭제
 const deleteStorelike = async (req: Request, res: Response) => {
-    const { mystoreId } = req.params;
+    const { MystoreId } = req.params;
     const userId = String(res.locals.user._id);
     const nickname = String(res.locals.user.nickname);
     try {
-        const findUser = await Favorite.findOne({ userId, mystoreId, nickname });
+        const findUser = await Favorite.findOne({ userId, MystoreId, nickname });
+        const Store = await Mystore.findById({ _id: MystoreId });
+        if (!Store) {
+            res.json({ result: false, msg: "존재하지 않는 값입니다." });
+            return;
+        }
         if (!findUser) {
             return res.json({ result: false, msg: "좋아요를 누르지 않았습니다." });
         } else {
             await Favorite.findByIdAndDelete(findUser);
+            let num: number = Store.favorite_count;
+            await myrecipe.findOneAndUpdate({ _id: MystoreId }, { $set: { favorite_count: --num } });
             return res.json({ result: true });
         }
     } catch (err) {
@@ -135,8 +159,7 @@ const deleteStorelike = async (req: Request, res: Response) => {
     }
 };
 
-
-//내가 좋아요 누른 마이레시피 조회
+//내가 좋아요 누른 스토어 조회
 const getMystore = async (req: Request, res: Response) => {
     const userId = String(res.locals.user._id);
     try {
@@ -148,7 +171,7 @@ const getMystore = async (req: Request, res: Response) => {
     }
 };
 
-export default { 
+export default {
     postlike,
     getAlluser,
     deletelike,
@@ -157,4 +180,4 @@ export default {
     getAllstoreuser,
     deleteStorelike,
     getMystore
- };
+};
